Asset Replacement (DONE)
Replaced as many platforms as I could with Conor's Assets. Some platforms are left greyboxed to be textured as shelves, replaced with dental floss box/string for rails, fan vents for propulsion and the green tower which will be replaced with piping that leads to bath taps for the bath which will fill the large empty area to the left of the starting area/south of the green tower.